refactor static notice to component, fixed footer links, change message notifications for ban and restriction action
