+++ app/components/common/static_notice/component.html.erb
@@ -7,6 +7,7 @@
       </div>
     </div>
   <% end %>
+
   <%= banned_banner %>
   <%= invalid_data_banner %>
 </div>
