+++ app/components/modals/change_offer_po/component.html.erb
@@ -23,7 +23,7 @@
           <div class="c-modal__grid">
             <div class="c-modal__main__info">
               <span><%= t('english_offers.show.min_bid') %></span>
-              <h3 class="u-h2 u-m-none"><%= number_with_precision(Money.new(Setting.find_by(code: 'auction_minimum_offer').retrieve),  precision: 2, separator: '.') %> €</h3>
+              <h3 class="u-h2 u-m-none"><%= price_format(Money.new(Setting.find_by(code: 'auction_minimum_offer').retrieve)) %></h3>
             </div>
             <div class="c-modal__main__info c-modal__main__info--black">
               <strong><%= t('auctions.you') %></strong>
