+++ config/locales/mailers/common_mailer.en.yml
@@ -5,3 +5,5 @@ en:
 
       <p>Respectfully,<br>
       Estonian Internet Foundation</p>
+
+  more_information_html: To participate in the auctions, go to <a href="%{link}">https://auction.internet.ee</a>.
