+++ config/locales/mailers/common_mailer.et.yml
@@ -7,3 +7,5 @@ et:
         <p>Lugupidamisega<br>
         Eesti Interneti Sihtasutus
         </p>
+
+  more_information_html: Oksjonitel osalemiseks minge aadressile <a href="%{link}">https://oksjon.internet.ee</a>.
